Experimental refactor of native-autofill.main.ts
